Update centeredpoint.el
Hi guys,

Great work! I have started using this minor mode and it has been useful. Initially, I searched for information on how to enable it globally. I found it elsewhere. I think this info could be in this repository somewhere. 

If this comment in this line (42) on this file is not the best place, maybe in the README file?

Thanks for sharing this.